refactor(errors): move domain exceptions to owners
