+++ src/frame_compare/analysis/frame_plan.py
@@ -6,7 +6,7 @@ import hashlib
 from dataclasses import dataclass
 from pathlib import Path
 
-from frame_compare.errors import InsufficientFramesError
+from frame_compare.analysis.errors import InsufficientFramesError
 
 __all__ = ["FramePlan", "select_uniform_seeded_frames", "create_frame_plan"]
