+++ src/frame_compare/orchestration/preparation.py
@@ -11,10 +11,8 @@ import frame_compare.analysis.cache_io as cache_io
 from frame_compare.config.overrides import apply_cli_overrides
 from frame_compare.config.schema import ConfigSchema
 from frame_compare.errors import (
-    AudioAlignmentError,
     CacheCorruptionError,
     CacheVersionMismatchError,
-    MetadataError,
     MetricsCalculationError,
 )
 from frame_compare.orchestration.context import (
@@ -39,6 +37,8 @@ from frame_compare.orchestration.types import (
 )
 from frame_compare.services.alignment import CACHE_FILE_NAME, check_alignment_cached
 from frame_compare.services.errors import (
+    AudioAlignmentError,
+    MetadataError,
     TmdbError,
     TmdbRateLimitedError,
 )
