+++ src/frame_compare/vs/errors.py
@@ -1,8 +1,10 @@
-"""VapourSynth subsystem error classes."""
+"""Video-source-specific exception types for Frame Compare."""
 
 from __future__ import annotations
 
-from frame_compare.errors import DependencyError, ErrorContext
+from pathlib import Path
+
+from frame_compare.errors import DependencyError, ErrorContext, ProcessingError
 
 
 class VapourSynthNotFoundError(DependencyError):
@@ -74,5 +76,35 @@ class TonemapRequiresVapourSynthError(DependencyError):
                 name="TONEMAP_REQUIRES_VAPOURSYNTH",
                 message="HDR tonemapping requires VapourSynth rendering",
                 hint="Disable [color].enable_tonemap for FFmpeg-only extraction, or use VapourSynth",
+            )
+        )
+
+
+class SourceLoadError(ProcessingError):
+    """Failed to initialize source engine (FC-4015)."""
+
+    def __init__(self, path: Path, engine_error: str) -> None:
+        super().__init__(
+            ErrorContext(
+                code="FC-4015",
+                name="SOURCE_LOAD_ERROR",
+                message=f"Failed to load source {path}: {engine_error}",
+                hint="Check file integrity or engine support",
+                details={"path": str(path), "engine_error": engine_error},
+            )
+        )
+
+
+class TonemapError(ProcessingError):
+    """Tonemapping failure (FC-4003)."""
+
+    def __init__(self, reason: str, hint: str | None = None) -> None:
+        super().__init__(
+            ErrorContext(
+                code="FC-4003",
+                name="TONEMAP_ERROR",
+                message=f"Tonemapping failed: {reason}",
+                hint=hint or "Check libplacebo support or config",
+                details={"reason": reason},
             )
         )
